rds_instance module - move rds import into recently agreed aws subdirectory
